+++ connpass/date.go
@@ -1,4 +1,4 @@
-package format
+package connpass
 
 import (
 	"fmt"
