+++ go.mod
@@ -1,11 +1,3 @@
-module github.com/lll-lll-lll-lll/conread
+module github.com/lll-lll-lll-lll/go-connpass
 
 go 1.18
-
-require github.com/stretchr/testify v1.8.0
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
